+++ dotnet/Eyes.Images.DotNet/Universal/Types/DebugImages.cs
@@ -2,11 +2,13 @@
 {
     public class DebugImages
     {
+        public bool Save { get; }
         public string Path { get; }
         public string Prefix { get; }
 
         public DebugImages(string path, string prefix)
         {
+            Save = true;
             Path = path;
             Prefix = prefix;
         }